refactor(scanner): extract strategy/vol.py — annualized_vol_yang_zhang (#225 PR4)
